+++ TimeTracker/Model/PEITimedTaskController.m
@@ -36,8 +36,8 @@
 
 - (void)initWithClientName:(NSString *)clientName
                        workSummary:(NSString *)workSummary
-                        hourlyRate:(int)hourlyrate
-                        timeWorked:(int)timeWorked {
+                        hourlyRate:(double)hourlyrate
+                        timeWorked:(double)timeWorked {
     
     //This should initialize a TimedTask
     PEITimedTask *newestTask = [[PEITimedTask alloc]
